work on time dependent bc
